# *: introduce image_pull_with_sync_fs in CRI