@fuweid 2023-11-21
It's based on the discussion at [weekly meeting](https://docs.google.com/document/d/1Q8KyVJd26oAQ3MafbnkVBgJCopPl2Bw45H9L9br9Vus/edit#heading=h.l5lxpvku4xjd) - September 14, 2023

ping @dmcgowan @mikebrow @samuelkarp @neersighted