@zhuangqh 2023-11-25
just writing file with O_direct?
write through to disk without page cache