@fuweid 2023-11-25
> just writing file with O_direct?

Hi @zhuangqh , there is not just regular file, but also symlink, block device and directory. For special files, it's unlikely to use O_direct to do persistence.