@mikebrow 2024-11-28
> Sorry to bother you. I used this script to set [image_pull_with_sync_fs](https://github.com/fuweid/go-dmflakey/blob/main/contrib/test/containerd/issue5854_test.go#L46) to true, but the final test result is still `/bin/busybox: exec format error`. Is there something wrong with my understanding of this MR?

What script? What MR? After changing the config.toml restart containerd. The test you linked is supposed to fail in order for it to pass.