@tboevil 2024-11-29
> What script? What MR? After changing the config.toml restart containerd. The test you linked is supposed to fail in order for it to pass.

Sorry, I didn't make it clear, I changed `ctrdCmd := runContainerd(t, rootfs, "", logPath)` to 

```golang
cfgPath := filepath.Join(t.TempDir(), "config.toml")
require.NoError(t,
    os.WriteFile(cfgPath, []byte(`
version = 2

[plugins]
  [plugins.'io.containerd.grpc.v1.cri']
    image_pull_with_sync_fs = true
`),
			0600),
	)
ctrdCmd := runContainerd(t, rootfs, cfgPath, logPath)
```

But the final execution result is still `bin/busybox: exec format error`. I thought that MR would ensure the write success after setting image_pull_with_sync_fs = true.