@tboevil 2024-11-29
```golang
func TestIssue5854(t *testing.T) {
	// NOTE: Set commit=1000 is to ensure that the global writeback won't
	// persist all the data during the simulation of power failure. So,
	// if the process doesn't call fsync/fdatasync, the data won't be
	// committed into disk.
	//
	// REF:
	// Query about ext4 commit interval vs dirty_expire_centisecs - https://lore.kernel.org/linux-ext4/20191213155912.GH15474@quack2.suse.cz/
	flakey := testutils.InitFlakeyDevice(t, t.Name(), dmflakey.FSTypeEXT4, "commit=1000")

	rootfs := flakey.RootFS()

	logPath := filepath.Join(t.TempDir(), "containerd.log")
	sockAddr := filepath.Join(rootfs, "run", "containerd", "containerd.sock")

	cfgPath := filepath.Join(t.TempDir(), "config.toml")
	require.NoError(t,
		os.WriteFile(cfgPath, []byte(`
version = 2

[plugins]
  [plugins.'io.containerd.grpc.v1.cri']
    image_pull_with_sync_fs = true
`),
			0600),
	)
	
	t.Log("Start to run containerd")
	ctrdCmd := runContainerd(t, rootfs, cfgPath, logPath)

	t.Log("Wait for ready")
	expectLog(t, logPath, "containerd successfully booted in")

	imageName := "ghcr.io/containerd/alpine:3.14.0"
	t.Logf("Pulling %s", imageName)
	t.Log(sockAddr)
	runCommand(t, "ctr", "-a", sockAddr, "i","pull", imageName)

	debugCmd := exec.Command("ctr", "-a", sockAddr, "i", "ls")
	out1, err := debugCmd.CombinedOutput()
	t.Logf("before: %s", out1)
	
	t.Log("Power failure")
	ctrdCmd.Process.Kill()
	require.Error(t, ctrdCmd.Wait())
	require.NoError(t, flakey.PowerFailure(""))

	t.Log("Restarting containerd")
	ctrdCmd = runContainerd(t, rootfs, "", logPath)

	t.Log("Wait for ready")
	expectLog(t, logPath, "containerd successfully booted in")
	defer ctrdCmd.Process.Kill()

	targetMount := filepath.Join(t.TempDir())
	defer sysutils.UnmountAll(targetMount, 0)

	t.Logf("Mounting image %s on %s", imageName, targetMount)
	debugCmd2 := exec.Command("ctr", "-a", sockAddr, "i", "ls")
	out2, err := debugCmd2.CombinedOutput()
	t.Logf("after: %s", out2)
	
	runCommand(t, "ctr", "-a", sockAddr, "image", "mount", imageName, targetMount)

	t.Logf("Run busybox (Chroot: %s)", targetMount)

	debugCmd3 := exec.Command("ls", "-al", path.Join(targetMount, "/bin"))
	out3, err := debugCmd3.CombinedOutput()
	t.Logf("ls -al `targetMount`: %s", out3)

	busyboxCmd := exec.Command(filepath.Join("/bin", "busybox"))
	busyboxCmd.SysProcAttr = &syscall.SysProcAttr{Chroot: targetMount}
	_, err = busyboxCmd.CombinedOutput()
	

	require.ErrorContains(t, err, "/bin/busybox: exec format error")
	t.Log(err.Error())
}
```

```shell
[root@cloudtower ~]# ./main -test.v
=== RUN   TestIssue5854
    issue5854_test.go:60: Start to run containerd
    issue5854_test.go:63: Wait for ready
    issue5854_test.go:67: Pulling ghcr.io/containerd/alpine:3.14.0
    issue5854_test.go:68: /tmp/TestIssue58541552633806/002/run/containerd/containerd.sock
    issue5854_test.go:73: before: REF                              TYPE                                                      DIGEST                                                                  SIZE    PLATFORMS                                                                                LABELS
        ghcr.io/containerd/alpine:3.14.0 application/vnd.docker.distribution.manifest.list.v2+json sha256:adab3844f497ab9171f070d4cae4114b5aec565ac772e2f2579405b78be67c96 2.7 MiB linux/386,linux/amd64,linux/arm/v6,linux/arm/v7,linux/arm64/v8,linux/ppc64le,linux/s390x -
    issue5854_test.go:75: Power failure
    issue5854_test.go:80: Restarting containerd
    issue5854_test.go:83: Wait for ready
    issue5854_test.go:90: Mounting image ghcr.io/containerd/alpine:3.14.0 on /tmp/TestIssue58541552633806/005
    issue5854_test.go:93: after: REF                              TYPE                                                      DIGEST                                                                  SIZE    PLATFORMS                                                                                LABELS
        ghcr.io/containerd/alpine:3.14.0 application/vnd.docker.distribution.manifest.list.v2+json sha256:adab3844f497ab9171f070d4cae4114b5aec565ac772e2f2579405b78be67c96 2.7 MiB linux/386,linux/amd64,linux/arm/v6,linux/arm/v7,linux/arm64/v8,linux/ppc64le,linux/s390x -
    issue5854_test.go:97: Run busybox (Chroot: /tmp/TestIssue58541552633806/005)
    issue5854_test.go:104: ls -al `targetMount`: total 8
        drwxr-xr-x  2 root root 4096 Jun 15  2021 .
        drwxr-xr-x 19 root root 4096 Nov 29 10:38 ..
        lrwxrwxrwx  1 root root   12 Jun 15  2021 arch -> /bin/busybox
        lrwxrwxrwx  1 root root   12 Jun 15  2021 ash -> /bin/busybox
        lrwxrwxrwx  1 root root   12 Jun 15  2021 base64 -> /bin/busybox
        lrwxrwxrwx  1 root root   12 Jun 15  2021 bbconfig -> /bin/busybox
        -rwxr-xr-x  1 root root    0 Jun 14  2021 busybox
        ......
    issue5854_test.go:109: fork/exec /bin/busybox: exec format error
--- PASS: TestIssue5854 (7.76s)
```

The complete code and results are shown above