@fuweid 2024-12-03
@tboevil Would you please show `containerd version`? just want to ensure that the `containerd` binary supports this flag thanks.