@tboevil 2024-12-04
> @tboevil Would you please show `containerd version`? just want to ensure that the `containerd` binary supports this flag thanks.

```bash
[root@cloudtower ~]# ctr -v
ctr containerd.io 1.7.23
[root@cloudtower ~]# runc -v
runc version 1.1.14
commit: v1.1.14-0-g2c9f560
spec: 1.0.2-dev
go: go1.22.8
libseccomp: 2.3.1
```